requirements.txt
psycopg2-binary
# psycopg2==2.8.6
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==6.2.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework
drf-yasg==1.21.3